headers/bsd: enhance compatibility with BSD
- Introduce conditional typedefs for timer_t
  and struct timespec to match BSD ABI layout.
- Adjust struct tm to use mutable __tm_zone for BSD
  while keeping const for Linux.
- Add BSD-specific constants (CLK_TCK, CLOCK_* macros, TIMER_RELTIME)
  aligned with system headers.
- Define BSD extensions in <grp.h> and unify setgroups() prototype
  across platforms (int vs size_t).
- Modify src/bsd/setgroups.c to use int count
  to match BSD system call interface.
- Add BSD-compatible futimes() implementation
  using SYS_futimes directly.
- Ensure proper isolation of Linux-specific fields
  and constants to prevent type or ABI mismatches.
- Improves safety by ensuring consistent data sizes
  and system call parameter types.

+++ src/bsd/setgroups.c
@@ -5,7 +5,7 @@
 #include "libc.h"
 
 struct ctx {
-	size_t count;
+	int count;
 	const gid_t *list;
 	int ret;
 };
@@ -26,7 +26,7 @@ static void do_setgroups(void *p)
 	c->ret = ret;
 }
 
-int setgroups(size_t count, const gid_t list[])
+int setgroups(int count, const gid_t list[])
 {
 	/* ret is initially nonzero so that failure of the first thread does not
 	 * trigger the safety kill above. */
